requirements.txt
asgiref
Django
djangorestframework
djoser
pytz
sqlparse